+++ src/model/othello/dto/Stone.java
@@ -1,6 +1,6 @@
-package model.dto;
+package model.othello.dto;
 
-import model.player.Player;
+import model.othello.player.Player;
 
 public final class Stone implements Cloneable {
   private Player player;
